Add some options to build script

+++ tensorflow/contrib/makefile/build_all_linux.sh
@@ -25,9 +25,12 @@ cd ${SCRIPT_DIR}/../../../
 source "${SCRIPT_DIR}/build_helper.subr"
 JOB_COUNT="${JOB_COUNT:-$(get_job_count)}"
 
-# Remove any old files first.
-make -f tensorflow/contrib/makefile/Makefile clean
-rm -rf tensorflow/contrib/makefile/downloads
+if [ -n $SHOULD_CLEAN ]; then
+    echo "Cleaning old files."
+    # Remove any old files first.
+    make -f tensorflow/contrib/makefile/Makefile clean
+    rm -rf tensorflow/contrib/makefile/downloads
+fi
 
 # Pull down the required versions of the frameworks we need.
 tensorflow/contrib/makefile/download_dependencies.sh
@@ -38,11 +41,23 @@ HOST_NSYNC_LIB=`tensorflow/contrib/makefile/compile_nsync.sh`
 TARGET_NSYNC_LIB="$HOST_NSYNC_LIB"
 export HOST_NSYNC_LIB TARGET_NSYNC_LIB
 
-# Compile protobuf.
-tensorflow/contrib/makefile/compile_linux_protobuf.sh
+if [ -z $USE_SYSTEM_PROTOBUF ]; then
+    echo "Using downloaded protobuf"
+    # Compile protobuf.
+    tensorflow/contrib/makefile/compile_linux_protobuf.sh
+else
+    echo "Using system protobuf"
+
+fi
+
+if [ -z $OPTFLAGS ]; then
+    OPTFLAGS="-O3 -march=native"
+fi
+
+echo "Optimization flags: $OPTFLAGS"
 
 # Build TensorFlow.
 make -j"${JOB_COUNT}" -f tensorflow/contrib/makefile/Makefile \
-  OPTFLAGS="-O3 -march=native" \
+  OPTFLAGS=$OPTFLAGS\
   HOST_CXXFLAGS="--std=c++11 -march=native" \
   MAKEFILE_DIR=$SCRIPT_DIR
